对接算法，修改错误的字段

+++ common/common-rpc-tools/src/main/java/com/iscas/biz/calculation/grpc/Sigma1Request.java
@@ -684,7 +684,7 @@ public  final class Sigma1Request extends
       output.writeUInt32NoTag(kuaChangMemoizedSerializedSize);
     }
     for (int i = 0; i < kuaChang_.size(); i++) {
-      output.writeDoubleNoTag(kuaChang_.get(i));
+      output.writeDoubleNoTag(kuaChang_.get(i).doubleValue());
     }
     if (girderDistance_ != 0D) {
       output.writeDouble(2, girderDistance_);
@@ -703,7 +703,7 @@ public  final class Sigma1Request extends
       output.writeUInt32NoTag(plateThickMemoizedSerializedSize);
     }
     for (int i = 0; i < plateThick_.size(); i++) {
-      output.writeDoubleNoTag(plateThick_.get(i));
+      output.writeDoubleNoTag(plateThick_.get(i).doubleValue());
     }
     if (deviceWeight_ != 0D) {
       output.writeDouble(7, deviceWeight_);
@@ -713,7 +713,7 @@ public  final class Sigma1Request extends
       output.writeUInt32NoTag(girderWidthMemoizedSerializedSize);
     }
     for (int i = 0; i < girderWidth_.size(); i++) {
-      output.writeDoubleNoTag(girderWidth_.get(i));
+      output.writeDoubleNoTag(girderWidth_.get(i).doubleValue());
     }
     if (materialYieldLimit_ != 0D) {
       output.writeDouble(9, materialYieldLimit_);
